Move RemoteMediaClient extensions
